Unify Scope and TypeScope, split into own module
